requirements.txt
Django>=4.2
openpyxl>=3.1.0
Pillow>=10.0.0
